Deprecate Proc::Async.start() in favor of ready()/finish()
Proc::Async has a start() method, which prepares a process to be
started.  After it returns, there is no guarantee that a process is
*actually* running, and some methods (for example: close-stdin, kill)
will silently fail, and more importantly, sometimes put the child
process and the VM's view of it in a state that locks up the program.

So start() is now deprecated; and it has changed so that after it
returns, the process is either running or has already exited.  If you
want to execute actions once a process has been started, you should use
ready:

    await $p.ready();
    # do stuff
    await $p.finish();